Sort every row starting from the 5th
I assume there's 4 rows of headers we don't want to sort. The number 5 is hardcoded into it all, but I figured it's easier to comprehend the code as a beginner this way.
This function is going to sort by the D column primarily, and by the E column secondarily. So if there's two rows that have the same values in the D column, the function will put them after each other, and sort these two locally by their value in the E column.
Lastly we number the rows, which wouldn't make sense before they're sorted.